pyproject.toml
[project]
name = "nli-by-gpt"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "kohei arasawa", email = "karasawa@lyria.inf.shizuoka.ac.jp" }
]
dependencies = [
    "openai>=1.55.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/nli_by_gpt"]
